Revert move of repeatCount (supported natively by skia)
