+++ coil-gif/src/androidMain/kotlin/coil3/gif/MovieDrawable.kt
@@ -283,6 +283,6 @@ class MovieDrawable @JvmOverloads constructor(
 
     companion object {
         /** Pass this to [setRepeatCount] to repeat infinitely. */
-        const val REPEAT_INFINITE = CONST_REPEAT_INFINITE
+        const val REPEAT_INFINITE = -1
     }
 }
